feat: cleanup components
